recieverPython/app/requirements.txt
flask==2.1.1
werkzeug==2.1.1
pika==1.2.1
Pillow==8.3.1
